Modificaciones en Skills parametro color y Corrección en Persona parametro id.

+++ src/main/java/com/portfolioweb/BackEndSpringBootAEG75/Controller/PersonaController.java
@@ -35,7 +35,7 @@ public class PersonaController {
    
    @PreAuthorize("hasRole('ADMIN')")
    @DeleteMapping("/personas/borrar/{id}")
-   public String deletePersona(@PathVariable Long id){
+   public String deletePersona(@PathVariable int id){
        ipersonaService.deletePersona(id);
        return "La persona fue eliminada correctamente";
    }
@@ -43,7 +43,7 @@ public class PersonaController {
    //URL:PUERTO/personas/editar/id/nombre & apellido & img
    @PreAuthorize("hasRole('ADMIN')")
    @PutMapping("/personas/editar/{id}")
-   public Persona editPersona(@PathVariable Long id,
+   public Persona editPersona(@PathVariable int id,
                               @RequestParam("nombre") String nuevoNombre,
                               @RequestParam("apellido") String nuevoApellido,
                               @RequestParam("img") String nuevoImg,
@@ -62,6 +62,6 @@ public class PersonaController {
 
    @GetMapping("/personas/traer/perfil")
    public Persona findPersona (){
-       return ipersonaService.findPersona((long)1);
+       return ipersonaService.findPersona((int)1);
    }
 }
