+++ src/main/java/com/portfolioweb/BackEndSpringBootAEG75/Interface/IPersonaService.java
@@ -11,8 +11,8 @@ public interface IPersonaService {
     public void savePersona(Persona persona);
     
     //Eliminar un usuario pero lo buscamos por ID
-    public void deletePersona(Long id);
+    public void deletePersona(int id);
     
     //Buscar una persona por ID
-    public Persona findPersona(Long id);   
+    public Persona findPersona(int id);   
 }
