+++ src/main/java/com/portfolioweb/BackEndSpringBootAEG75/Service/ImpPersonaService.java
@@ -23,12 +23,12 @@ public class ImpPersonaService implements IPersonaService{
     }
 
     @Override
-    public void deletePersona(Long id) {
+    public void deletePersona(int id) {
         ipersonaRepository.deleteById(id);
     }
 
     @Override
-    public Persona findPersona(Long id) {
+    public Persona findPersona(int id) {
         Persona persona = ipersonaRepository.findById(id).orElse(null);
         return persona;
     }
